requirements.txt
_libgcc_mutex==0.1
ca-certificates==2021.4.13
certifi==2020.12.5
grpcio-tools==1.37.0
grpcio==1.37.0
ld_impl_linux-64==2.33.1
libffi==3.3
libgcc-ng==9.1.0
libstdcxx-ng==9.1.0
ncurses==6.2
numpy==1.20.2
openssl==1.1.1k
Pillow==8.2.0
pip==21.0.1
protobuf==3.15.8
python==3.9.4
readline==8.1
setuptools==52.0.0
six==1.15.0
sqlite==3.35.4
tk==8.6.10
tzdata==2020f
wheel==0.36.2
xz==5.2.5
zlib==1.2.11
